connected bootcamp with users and only the creator of the bootcamp can update or delete. Publishers can only create one bootcamp and admin functionality is done.

+++ controllers/bootcampControllers.js
@@ -28,6 +28,17 @@ export const getBootcamp = asyncHandler(async (req, res, next) => {
 // @route   POST /api/v1/bootcamps
 // @access  Private
 export const createBootcamp = asyncHandler(async (req, res, next) => {
+    // Add user to req.body
+    req.body.user = req.user.id
+
+    // Check for published bootcamp
+    const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id })
+
+    // If the user is not an admin, they can only add one bootcamp
+    if (publishedBootcamp && req.user.role !== "admin") {
+        return next(new ErrorResponse(`The use with ID ${req.user.id} has already published a bootcamp`), 400)
+    }
+
     const bootcamp = await Bootcamp.create(req.body)
     res.status(StatusCodes.CREATED).json({ success: true, data: bootcamp })
 })
@@ -37,14 +48,22 @@ export const createBootcamp = asyncHandler(async (req, res, next) => {
 // @access  Private
 export const updateBootcamp = asyncHandler(async (req, res, next) => {
     const { id } = req.params
-    const updatedBootcamp = await Bootcamp.findByIdAndUpdate(id, req.body, {
+    let bootcamp = await Bootcamp.findById(id)
+    if (bootcamp) {
+        return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404))
+    }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401))
+    }
+
+    bootcamp = await Bootcamp.findByIdAndUpdate(id, req.body, {
         new: true,
         runValidators: true,
     })
-    if (!updateBootcamp) {
-        return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404))
-    }
-    res.status(StatusCodes.OK).json({ success: true, data: updatedBootcamp })
+
+    res.status(StatusCodes.OK).json({ success: true, data: bootcamp })
 })
 
 // @desc    Delete bootcamp
@@ -56,6 +75,12 @@ export const deleteBootcamp = asyncHandler(async (req, res, next) => {
     if (!removedBootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404))
     }
+
+    // Make sure user is bootcamp owner
+    if (removedBootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.params.id} is not authorized to delete this bootcamp`, 401))
+    }
+
     await removedBootcamp.deleteOne()
     res.status(StatusCodes.OK).json({ success: true, data: removedBootcamp })
 })
@@ -95,6 +120,12 @@ export const bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
     if (!bootcamp) {
         return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404))
     }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`, 401))
+    }
+
     if (!req.files) {
         return next(new ErrorResponse(`Please upload a file`, 400))
     }
